pyproject.toml
[project]
name = "talkbot"
version = "0.1.0"
description = "A talking bot demo using OpenRouter and pyttsx3"
readme = "README.md"
authors = [
    { name = "ricklon", email = "rick.rickanderson@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "edge-tts>=6.1",
    "httpx>=0.24",
    "kittentts @ https://github.com/KittenML/KittenTTS/releases/download/0.8/kittentts-0.8.0-py3-none-any.whl",
    "pygame>=2.5",
    "pyttsx3>=2.90",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[project.scripts]
talkbot = "talkbot.cli:main"
talkbot-gui = "talkbot.gui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/talkbot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that exercise real external/runtime backends",
]
